+++ R/plot1GS.R
@@ -132,6 +132,12 @@
 #'numeric value.  It specifies the function used to aggregate the observations
 #'before the clustering.  Default is to \code{"mean"}.
 #'
+#'@param na.rm.aggreg
+#'a logical flag indicating whether \code{NA} should be remove to prevent 
+#'propagation through \code{aggreg.fun}. Can be useful to set to TRUE with 
+#'unbalanced design as those will generate structural \code{NA}s in 
+#'\code{$Estimations}. Default is \code{TRUE}.
+#'
 #'@param trend.fun 
 #'a character string such as \code{"mean"} or
 #'the name of any other function that returns a single numeric value.  It
@@ -341,7 +347,8 @@ plot1GS <-
 			 baseline=NULL,
 			 group.var=NULL, Group_ID_paired=NULL, ref=NULL, group_of_interest=NULL,
 			 FUNcluster=NULL, clustering_metric="euclidian", clustering_method="ward", B=500,
-			 max_trends=4, aggreg.fun="median", trend.fun="median",
+			 max_trends=4, aggreg.fun="median", na.rm.aggreg=TRUE,
+			 trend.fun="median",
 			 methodOptiClust = "firstSEmax",
 			 indiv="genes",
 			 verbose=TRUE,
@@ -353,7 +360,15 @@ plot1GS <-
 			 plot=TRUE
 	){
 		
+		if(!is.null(group.var) | !is.null(ref) | !is.null(group_of_interest)){
+			warning("multigroup option is not implemented yet for plot1GS\n Defaulting back to only 1 group representation instead")	
+		}
+		if(is.null(group.var) & (!is.null(group_of_interest) | !is.null(ref))){
+			stop("'group.var' is NULL while 'group_of_interest' or 'ref' is not")
+		}
+		
 		pre_clustering <- !is.null(precluster)
+		clustering <- !pre_clustering
 		
 		capwords <- function(s, strict = FALSE){
 			cap <- function(s){
@@ -433,9 +448,9 @@ plot1GS <-
 
 		data_stand <- t(apply(X=data_sel, MARGIN=1, FUN=scale))
 		if(indiv == "genes"){
-			data_stand_MedianByTP <- t(apply(X=data_stand, MARGIN=1, FUN=Fun_byIndex, index=as.factor(TimePoint), fun=aggreg.fun, na.rm=TRUE))
+			data_stand_MedianByTP <- t(apply(X=data_stand, MARGIN=1, FUN=Fun_byIndex, index=as.factor(TimePoint), fun=aggreg.fun, na.rm=na.rm.aggreg))
 		}else if(indiv=="patients"){
-			data_tocast<-cbind.data.frame(TimePoint, Subject_ID, "M" = apply(X=data_stand, MARGIN=2, FUN=aggreg.fun, na.rm=TRUE))
+			data_tocast <- cbind.data.frame(TimePoint, Subject_ID, "M" = apply(X=data_stand, MARGIN=2, FUN=aggreg.fun, na.rm=na.rm.aggreg))
 			data_stand_MedianByTP <- as.matrix(acast(data_tocast, formula="Subject_ID~TimePoint", value.var="M"))
 		}
 		
@@ -488,14 +503,18 @@ plot1GS <-
 			}else{
 				
 				clust <- precluster
+				if(indiv=="genes"){
 				medoids <- as.data.frame(t(apply(X=data_stand_MedianByTP, MARGIN=2, FUN=Fun_byIndex, 
 												 index=as.factor(as.numeric(precluster)), fun=trend.fun, na.rm=TRUE)))
+				}else if(indiv=="patients"){
+					stop("'precluster' is only implemented for indiv = 'genes'")
+				}
 				if(nrow(medoids)==1){
 					medoids <- cbind.data.frame("TimePoint"= colnames(medoids), "1"=t(medoids))
 				}else{
 					medoids <- cbind.data.frame("TimePoint"= rownames(medoids), medoids)
 				}
-				colnames(medoids) <- c("TimePoint", levels(precluster))
+				colnames(medoids) <- c("TimePoint", levels(as.factor(as.numeric(precluster))))
 			}
 			if(verbose){
 				message("DONE\n")
@@ -518,6 +537,7 @@ plot1GS <-
 			medoids <- cbind.data.frame("TimePoint"=colnames(data_stand_MedianByTP), "1"='NA')
 			clust <- rep(NA, dim(data_stand_MedianByTP)[1])
 		}
+		
 		
 		medoids$TimePoint <- as.numeric(as.character(medoids$TimePoint))
 		colnames(data_stand_MedianByTP) <- as.numeric(colnames(data_stand_MedianByTP))
@@ -606,7 +626,6 @@ plot1GS <-
 			  + theme(plot.margin=unit(margins*c(0.5, 0.7, 0.1, 0.5), 'lines'))
 			  + theme(legend.key=element_rect(fill="white"))
 		)
-		
 		
 		if(showTrend){