+++ inst/doc/TcGSA_userguide.R
@@ -1,10 +1,13 @@
 ## ----pre, echo=FALSE, warning=FALSE, include=FALSE-----------------------
 library(knitr)
 is_check <- ("CheckExEnv" %in% search()) || any(c("_R_CHECK_TIMINGS_", "_R_CHECK_LICENSE_") %in% names(Sys.getenv()))
-knitr::opts_chunk$set(echo = TRUE, eval=!is_check, cache=TRUE)
+knitr::opts_chunk$set(echo = TRUE, eval=!is_check, cache=TRUE,
+					  collapse = TRUE,
+					  comment = "#>"
+)
 #rmarkdown::render("vignettes/TcGSA_userguide.Rmd")
 
-## ----GS_import, message=FALSE, cache=TRUE--------------------------------
+## ----GS_import, message=FALSE--------------------------------------------
 temp <- tempfile()
 utils::download.file("http://doi.org/10.1371/journal.pcbi.1004310.s007", destfile = temp, mode = "wb")
 load(unz(temp, "ReproducibleRFiles/GMTs_PLOScb.RData", open = "r"))
@@ -19,10 +22,10 @@ if (!requireNamespace("GEOquery", quietly = TRUE)) {
 	BiocManager::install("GEOquery")
 }
 
-## ----import_dalia, message=FALSE, cache=TRUE, results='hide'-------------
+## ----import_dalia, message=FALSE, results='hide'-------------------------
 GEOquery::getGEOSuppFiles('GSE46734', filter_regex="(*NonParamCombat*)|(*DESIGN*)")
 
-## ----design_dalia, cache=TRUE--------------------------------------------
+## ----design_dalia--------------------------------------------------------
 design <- read.delim(gzfile("GSE46734/GSE46734_DALIA1longitudinalTranscriptome_DESIGN_anonym.txt.gz"))
 design_preATI <- design[-which(design$TimePoint<0 | design$TimePoint==16 | design$TimePoint>22), ]
 head(design_preATI,5)
@@ -43,26 +46,26 @@ stopifnot(nrow(expr_preATI)==32978 & ncol(expr_preATI)==90)
 ## ----dim_expr_DALIA------------------------------------------------------
 identical(ncol(expr_preATI), nrow(design_preATI))
 
-## ----LR_ST, message=FALSE, warning=FALSE, cache=TRUE---------------------
+## ----LR_ST, message=FALSE, warning=FALSE---------------------------------
 tcgsa_result <- TcGSA::TcGSA.LR(expr = expr_preATI, 
 								gmt = gmt_modulesV2, 
 								design = design_preATI, 
 								subject_name = "Patient_ID", 
 								time_name = "TimePoint")
 
-## ----tcgsa_result, cache=TRUE, echo=FALSE--------------------------------
+## ----tcgsa_result, echo=FALSE--------------------------------------------
 tcgsa_result
 
-## ----summary_dalia, cache=TRUE-------------------------------------------
+## ----summary_dalia-------------------------------------------------------
 summary(tcgsa_result)
 
-## ----signifLRT_ST, cache=TRUE--------------------------------------------
+## ----signifLRT_ST--------------------------------------------------------
 head(TcGSA::signifLRT.TcGSA(tcgsa_result)$mixedLRTadjRes)
 
-## ----multtest_ST, cache=TRUE---------------------------------------------
+## ----multtest_ST---------------------------------------------------------
 head(TcGSA::multtest.TcGSA(tcgsa_result))
 
-## ----plot1GS_ST, message=FALSE, warning=FALSE, fig.keep='all', cache=TRUE----
+## ----plot1GS_ST, message=FALSE, warning=FALSE, fig.keep='all'------------
 TcGSA::plot1GS(expr = expr_preATI, 
 			   #plot1GS(expr = tcgsa_result$Estimations,
 			   gmt = gmt_modulesV2, 
@@ -80,11 +83,11 @@ TcGSA::plot1GS(expr = expr_preATI,
 			   			ggplot2::ylim(-1.26,1.26)
 			   ))
 
-## ----import_ober, message=FALSE, warning=FALSE, cache=TRUE---------------
+## ----import_ober, message=FALSE, warning=FALSE---------------------------
 utils::download.file("ftp://ftp.ncbi.nlm.nih.gov/geo/series/GSE30nnn/GSE30101/soft/GSE30101_family.soft.gz", destfile = "GSE30101_family.soft.gz", mode = "wb", cacheOK = FALSE)
 gse.soft <- GEOquery::getGEO(filename="GSE30101_family.soft.gz")
 
-## ----expr_ober, cache=TRUE-----------------------------------------------
+## ----expr_ober-----------------------------------------------------------
 probesIDs <- GEOquery::Table(GEOquery::GSMList(gse.soft)[[1]])$ID
 data.matrix <- do.call('cbind', lapply(GEOquery::GSMList(gse.soft),
 									   function(x) {
@@ -96,7 +99,7 @@ rownames(data.matrix) <- probesIDs
 expr.All.ChaussVac <- apply(X = data.matrix, MARGIN = 2, FUN = as.numeric)
 rownames(expr.All.ChaussVac) <- probesIDs
 
-## ----design_ober, cache=TRUE---------------------------------------------
+## ----design_ober---------------------------------------------------------
 design_list <- lapply(GEOquery::GSMList(gse.soft), 
 					  function(x){GEOquery::Meta(x)$characteristics_ch1})
 design <- data.frame(row.names = names(design_list))
@@ -172,7 +175,7 @@ design.PNEUMOvsSALINE.ChaussVac.AvgBl$Vaccine <- as.factor(as.character(design.P
 
 expr.PNEUMOvsSALINE.ChaussVac.AvgBl <- expr.All.ChaussVac.AvgBl[,design.PNEUMOvsSALINE.ChaussVac.AvgBl$sample_ID]
 
-## ----LR_MT, cache=TRUE, message=FALSE, warning=FALSE---------------------
+## ----LR_MT, message=FALSE, warning=FALSE---------------------------------
 tcgsa_result_MT <- TcGSA::TcGSA.LR(expr = expr.PNEUMOvsSALINE.ChaussVac.AvgBl, 
 								   gmt = gmt_modulesV2,
 								   design = design.PNEUMOvsSALINE.ChaussVac.AvgBl, 
@@ -182,19 +185,20 @@ tcgsa_result_MT <- TcGSA::TcGSA.LR(expr = expr.PNEUMOvsSALINE.ChaussVac.AvgBl,
 summary(tcgsa_result_MT)
 head(TcGSA::signifLRT.TcGSA(tcgsa_result_MT)$mixedLRTadjRes)
 
-## ----clust_MT, message=FALSE, warning=FALSE, cache=TRUE------------------
+## ----clust_MT, message=FALSE, warning=FALSE------------------------------
 clust <- TcGSA::clustTrend(tcgs = tcgsa_result_MT, 
 						   expr=tcgsa_result_MT$Estimations,
-						   Subject_ID=design_gen$Patient_ID,
-						   TimePoint=design_gen$TimePoint,
-						   baseline = 0,
-						   group_of_interest="pneumo")
+						   Subject_ID=design.PNEUMOvsSALINE.ChaussVac.AvgBl$Patient_ID,
+						   TimePoint=design.PNEUMOvsSALINE.ChaussVac.AvgBl$Day,
+						   group.var = design.PNEUMOvsSALINE.ChaussVac.AvgBl$Vaccine,
+						   group_of_interest="pneumo",
+						   ref="saline")
 clust
 
-## ----heatmap_MT, message=FALSE, cache=TRUE, results='asis'---------------
+## ----heatmap_MT, message=FALSE, results='asis'---------------------------
 plot(x=tcgsa_result_MT, expr=tcgsa_result_MT$Estimations,
-	 Subject_ID=design_gen$Patient_ID,
-	 TimePoint=design_gen$TimePoint,
+	 Subject_ID=design.PNEUMOvsSALINE.ChaussVac.AvgBl$Patient_ID,
+	 TimePoint=design.PNEUMOvsSALINE.ChaussVac.AvgBl$TimePoint,
 	 group_of_interest="pneumo",
 	 clust_trends=clust,
 	 legend.breaks=seq(from=-2,to=2, by=0.01), time_unit="D",