+++ NEWS.md
@@ -1,5 +1,9 @@
 # News about the `TcGSA` R package
 
+### Main changes in Version 0.12.4 (2019-04-30) --- *this is only a minor release*:
+* Prevent the simultaneous use of `baseline` and `group.var` in plotting functions 
+* Added `na.rm`support for `aggreg.fun` everywhere
+
 ### Main changes in Version 0.11.0 (2018-06-07) --- *this is only a minor release*:
 * Added a vignette
 * p-values are no-longer simulated but computed from asymptotic chi^2 mixture